requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
google-api-python-client==2.86.0
flask-cors==4.0.0
flask-limiter==3.3.0
httplib2==0.22.0
requests==2.31.0
jinja2==3.0.1
itsdangerous==2.0.1
markupsafe==2.0.1